+++ src/module.ts
@@ -21,6 +21,8 @@ export default defineNuxtModule<ModuleOptions>({
 
     nuxt.options.css.push('mapbox-gl/dist/mapbox-gl.css')
     nuxt.options.app.head.link?.push({rel: 'stylesheet', href: 'https:/api.mapbox.com/mapbox-gl-js/plugins/mapbox-gl-geocoder/v5.0.0/mapbox-gl-geocoder.css', type: 'text/css'})
+    nuxt.options.app.head.script?.push({ src: 'https://api.mapbox.com/mapbox-gl-js/v2.12.0/mapbox-gl.js' })
+    nuxt.options.app.head.script?.push({ src: 'https://api.mapbox.com/mapbox-gl-js/plugins/mapbox-gl-geocoder/v5.0.0/mapbox-gl-geocoder.min.js' })
 
     const appConfig = nuxt.options.appConfig as ExtendedAppConfig
     appConfig._MAPBOX_API_KEY = options.apiKey;