+++ src/runtime/components/Geocoder.vue
@@ -1,12 +1,12 @@
 <!-- eslint-disable vue/multi-word-component-names -->
 <script setup lang="ts">
-    import MapboxGeocoder from '@mapbox/mapbox-gl-geocoder';
-    import mapboxgl from 'mapbox-gl';
+    import {GeocoderOptions} from '@mapbox/mapbox-gl-geocoder';
+    // import mapboxgl from 'mapbox-gl';
     import { inject } from 'vue';
     import { useMapbox } from '../composables/useMapbox';
 
     interface Props {
-        options: Omit<MapboxGeocoder.GeocoderOptions, "accessToken">
+        options: Omit<GeocoderOptions, "accessToken">
     }
     const props = defineProps<Props>();
 
@@ -16,7 +16,8 @@
 
     useMapbox(mapId, (map) => {
         function addControl(){
-            map?.addControl(new MapboxGeocoder({accessToken: mapboxgl.accessToken, ...props.options}))
+          //@ts-ignore TODO: Figure out typing while getting around #2
+          map?.addControl(new MapboxGeocoder({accessToken: mapboxgl.accessToken, mapboxgl, ...props.options}))
         }
 
         map.on('load', addControl)