+++ src/runtime/plugin.client.ts
@@ -1,17 +1,17 @@
 import { ExtendedAppConfig } from '../module';
 import { defineNuxtPlugin, useAppConfig, useState } from '#app'
-import mapbox from "mapbox-gl"
-import mapboxgl from 'mapbox-gl';
+// import mapboxgl from 'mapbox-gl'
+// import { Map, MapboxOptions, PopupOptions, Popup, MarkerOptions, Marker } from 'mapbox-gl';
 import { Ref } from 'vue';
 
 export interface MapboxInstancesObject {
-    [key: string]: mapbox.Map
+    [key: string]: mapboxgl.Map
 }
 export interface MapboxPopupsObject {
-  [key: string]: mapbox.Popup
+  [key: string]: mapboxgl.Popup
 }
 export interface MapboxMarkerObject {
-  [key: string]: mapbox.Marker
+  [key: string]: mapboxgl.Marker
 }
 
 export default defineNuxtPlugin((nuxtApp) => {
@@ -20,24 +20,27 @@ export default defineNuxtPlugin((nuxtApp) => {
     const mapbox_marker_instances: Ref<MapboxMarkerObject> = useState('mapbox_marker_instances', () => {return {}})
 
     const appConfig = useAppConfig() as ExtendedAppConfig
-    
+    //@ts-ignore
     mapboxgl.accessToken = appConfig._MAPBOX_API_KEY;
 
 
     return {
       provide: {
-          mapboxInit: (key: string, options: mapbox.MapboxOptions = {container: key}) => {
-            mapbox_instances.value[key] = new mapbox.Map(options)
+          mapboxInit: (key: string, options: mapboxgl.MapboxOptions = {container: key}) => {
+            //@ts-ignore
+            mapbox_instances.value[key] = new mapboxgl.Map(options)
           },
           mapboxInstances: () => mapbox_instances,
 
-          mapboxInitPopup: (key: string, options: mapbox.PopupOptions) => {
-            mapbox_popup_instances.value[key] = new mapbox.Popup(options)
+          mapboxInitPopup: (key: string, options: mapboxgl.PopupOptions) => {
+            //@ts-ignore
+            mapbox_popup_instances.value[key] = new mapboxgl.Popup(options)
           },
           mapboxPopupInstances: () => mapbox_popup_instances,
 
-          mapboxInitMarker: (key: string, options: mapbox.MarkerOptions) => {
-            mapbox_marker_instances.value[key] = new mapbox.Marker(options)
+          mapboxInitMarker: (key: string, options: mapboxgl.MarkerOptions) => {
+            //@ts-ignore
+            mapbox_marker_instances.value[key] = new mapboxgl.Marker(options)
           },
           mapboxMarkerInstances: () => mapbox_marker_instances,
       }