+++ CHANGELOG.md
@@ -7,8 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Popup Control
+- More config options
+## v1.1.7
+## Fixed
+- Fixed `Map markers are disabled.` with geocoder
 
+## v1.1.6
+## Fixed
+- Import fix
+
+## v1.1.0
+### Added
+
+- Default Popup & Marker components
+- defineMapboxMarker & defineMapboxPopup for easily making custom components
+
+### v1.0.1
 ### Added
 
 - New syntax for useMapbox with callback